aes_lst

set aesthetics as a named list for: 1) mapping i) estimated locations; ii) estimated confidence ellipses; iii) estimated track lines; iv) observed locations; v) land masses; vi) water bodies; and 2) colour palettes for i) behavioural indices or ii) individual animal tracks

aes_lst(
  est = TRUE,
  conf = TRUE,
  line = FALSE,
  mp = TRUE,
  obs = FALSE,
  shape = c(19, NA, NA, 17, NA, NA),
  size = c(1.25, NA, 0.2, 0.8, NA, NA),
  col = c("dodgerblue", NA, "grey50", "orange", NA, NA),
  fill = c("dodgerblue", "dodgerblue", NA, "orange", "grey60", "grey85"),
  alpha = c(1, 0.4, 1, 1, 1, NA),
  mp_pal = hcl.colors(100, palette = "Plasma", rev = FALSE),
  id_pal = "Harmonic",
  date_pal = hcl.colors(100, palette = "Viridis", rev = FALSE)
)

Arguments

est

logical; turns on estimated locations (default = TRUE)

conf

logical; turns on estimated confidence ellipses. Default varies depending on whether behavioural index is being mapped &/or if single versus multiple tracks are being mapped.

line

logical; turns on estimated track line(s) (default varies)

mp

logical; turns on move persistence index (default = TRUE, if present in model fit)

obs

logical; turns on observed locations (default = FALSE)

shape

gpplot2 shape value (integer: 0, 25) for estimated & observed locations

size

ggplot2 size value for estimated locations & track lines, and observed locations

col

colour for estimated locations and track lines, and observed locations

fill

fill colour for estimated locations & confidence ellipses, observed locations, land polygons, and water

alpha

transparency for specified fills/colours

mp_pal

continuous colour palette for move persistence values

id_pal

discrete colour palette for track id's

date_pal

continuous colour palette for displaying date along track

Value

a named list, with 9 elements, of map components and aesthetics

elements 1-5 - map components: est, conf, line, mp, obs

element 6: a data.frame named df containing ggplot2 aesthetics: shape, size, col, fill, and alpha. df has 6 rows, corresponding to the map features: estimated locations, confidence ellipses, track lines, observed locations, land polygons, and water

element 7: mp_pal

element 8: id_pal

element 9: date_pal

Examples

# generate custom aes list
aes <- aes_lst(conf = FALSE, mp_pal = hcl.colors(n=100, palette = "RdBu"))

# modify aesthetics
aes$df$size[1] <- 1.5
aes$df$fill[6] <- grey(0.9)