fit a Move Persistence Model (mpm)

fit a random walk with time-varying move persistence to temporally regular or irregular location data

fit_mpm(
  x,
  what = "predicted",
  model = c("jmpm", "mpm"),
  coords = 3:4,
  control = mpm_control(),
  inner.control = NULL,
  optim = NULL,
  optMeth = NULL,
  verbose = NULL
)

Arguments

x

a ssm_df fit object or a data frame of observations (see details)

what

if a ssm_df fit object is supplied then what determines whether fitted, predicted (default), or rerouted values are mapped; ignored if x is a data frame

model

mpm model to fit; either mpm with unpooled random walk variance parameters (sigma_(g,i)) or jmpm with a single, pooled random variance parameter (sigma_g)

coords

column numbers of the location coordinates (default = 3:4)

control

list of control settings for the outer optimizer (see mpm_control for details)

inner.control

list of control parameters for the inner optimization

optim

is deprecated, use ssm_control(optim = "optim") instead, see ssm_control for details

optMeth

is deprecated, use ssm_control(method = "L-BFGS-B") instead, see ssm_control for details

verbose

is deprecated, use ssm_control(verbose = 1) instead, see ssm_control for details

Value

a list with components

fitted a dataframe of fitted locations

par model parameter summary

data input data.frame

tmb the TMB object

opt the object returned by the optimizer

References

Jonsen ID, McMahon CR, Patterson TA, et al. (2019) Movement responses to environment: fast inference of variation among southern elephant seals with a mixed effects model. Ecology. 100(1):e02566

Examples

## fit jmpm to two southern elephant seal tracks
xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72, control = ssm_control(verbose = 0))
#> 
#> 

fmpm <- fit_mpm(xs, model = "jmpm")
#> fitting jmpm...
#>  pars:   0 0 0       pars:   -0.63038 -0.77416 0.05745       pars:   -0.21961 -0.90381 0.95993       pars:   0.1024 -0.34798 0.19353       pars:   -0.12912 -0.74761 0.74455       pars:   -0.17853 -0.8329 0.86216       pars:   -0.23363 -0.801 0.86597       pars:   -0.27635 -0.61987 0.76146       pars:   -0.40458 -0.79384 0.62587       pars:   -0.36412 -0.6719 0.40545       pars:   -0.38583 -0.70377 0.60569       pars:   -0.43697 -0.67572 0.577       pars:   -0.38583 -0.70377 0.60569