fmap

map aniMotum fitted or predicted locations, with or without Argos observations, optionally apply a different projection

fmap(
  x,
  y = NULL,
  what = c("fitted", "predicted", "rerouted"),
  conf = TRUE,
  obs = FALSE,
  obs.shp = 17,
  by.id = TRUE,
  by.date = TRUE,
  crs = NULL,
  ext.rng = c(0.05, 0.05),
  size = 0.25,
  col.ssm = "dodgerblue",
  col.obs = "black",
  lines = FALSE,
  landfill = grey(0.6),
  map_type = "default",
  pal = "Cividis",
  rev = FALSE,
  last_loc = NULL,
  ...
)

Arguments

x

a aniMotum ssm fit object with class ssm_df

y

optionally, a aniMotum mpm fit object with class mpm_df; default is NULL

what

specify which location estimates to map: fitted, predicted or rerouted

conf

include confidence regions around estimated location (logical; default = TRUE, unless y is an mpm fit object then conf is FALSE)

obs

include Argos observations on map (logical; default = FALSE)

obs.shp

point shape for observations (default = 17)

by.id

when mapping multiple tracks, should locations be coloured by id (logical; default = TRUE if nrow(x) > 1 else FALSE)

by.date

when mapping single tracks, should locations be coloured by date (logical; default = TRUE if nrow(x) == 1 else FALSE)

crs

proj4string for re-projecting locations, if NULL the default projection ("+proj=merc") for the fitting the SSM will be used

ext.rng

factors to extend the plot range in x and y dimensions (can exceed 1)

size

size of estimated location points (size = NA will draw no points). Optionally, a vector of length 2 with size of observed locations given by 2nd value (ignored if obs = FALSE)

col.ssm

colour of ssm-fitted or ssm-predicted locations (ignored if by.id = TRUE)

col.obs

colour of observed locations (ignored if obs = FALSE)

lines

logical indicating if lines are added to connect estimated locations (default = FALSE)

landfill

colour to use for land (default = grey(0.6))

map_type

background map type (default = NULL, which uses rnaturalearth to add landmasses); if packages ggspatial & rosm are installed then any map type returned by rosm::osm.types can be used for a more detailed map background.

pal

hcl.colors palette to use (default: "Cividis"; type hcl.pals() for options)

rev

reverse colour palette (logical)

last_loc

colour to render last location of each track (default = NULL)

...

additional arguments passed to ggspatial::annotation_map_tile