Coerce input data into expected aniMotum format

format data by mapping supplied variable names to those expected by fit_ssm(), and ensuring variables are put into the expected order. Can be run manually by user as a data pre-processing step prior to calling fit_ssm() or can be called automatically by fit_ssm(). In the latter case, any custom variable names must be declared as arguments to fit_ssm(); see examples, below.

format_data(
  x,
  id = "id",
  date = "date",
  lc = "lc",
  coord = c("lon", "lat"),
  epar = c("smaj", "smin", "eor"),
  sderr = c("x.sd", "y.sd"),
  tz = "UTC"
)

Arguments

x

input data

id

the name (as a quoted character string) of id variable: a unique identifier for individual (animal) track data sets.

date

the name (as a quoted character string)of the date/time variable: date and time (as YYYY-MM-DD HH:MM:SS) of each observation.

lc

the name (as a quoted character string) of the location quality class variable: Argos location quality class (values in the set: 3,2,1,0,"A","B","Z"). Can also include "G" for GPS data and/or "GL" for light-level geolocation (GLS) and other data types.

coord

the names (as quoted character strings) of the location coordinate variables: defaults are c("lon","lat"), but could also be c("x","y") for planar coordinates; or if input data is an sf object then "geometry". If input data is an sf object then coord is set to "geometry" by default.

epar

the names (as quoted character strings) of the Argos error ellipse parameters: defaults are "smaj" (ellipse semi-major axis), "smin" (ellipse semi-minor axis), and "eor" (ellipse orientation). Ignored if these variables are missing from the input data.

sderr

the names (as quoted character strings) of provided standard errors for lon,lat or x,y: default names are x.sd, y.sd. Typically, these are only provided for generic location data such as processed light-level geolocations, or high-resolution acoustic detections. The argument is ignored if these variables are missing from the input data.

tz

the timezone the applies to the data/time variable if they are not in tz = 'UTC'. A list of valid timezone names can be viewed via OlsonNames()

Value

a data.frame or sf-tibble of input data in expected aniMotum format. Additional columns required by fit_ssm(), if missing, will be added to the formatted tibble: smaj, smin, eor, x.sd, and y.sd.

Examples

## as a data pre-processing step
data(sese2_n)
head(sese2_n, 5)
#> # A tibble: 5 × 5
#>   longitude latitude time                lc    id       
#>       <dbl>    <dbl> <chr>               <fct> <chr>    
#> 1      72.5    -50.2 2009-02-01 17:50:46 A     ct36-E-09
#> 2      73.0    -50.4 2009-02-02 03:30:26 A     ct36-E-09
#> 3      73.8    -50.8 2009-02-02 17:50:48 B     ct36-E-09
#> 4      74.6    -51.2 2009-02-03 07:39:08 A     ct36-E-09
#> 5      74.9    -51.4 2009-02-03 15:29:59 A     ct36-E-09
d <- format_data(sese2_n, date = "time", coord = c("longitude","latitude"), 
tz = "America/Halifax")
fit <- fit_ssm(d, model = "crw", time.step = 24)
#> fitting crw SSM to 2 tracks...
#>  pars:   1 1 0 0 -1.82873 0       pars:   0.593 0.46862 -0.38936 -0.59345 -2.04237 0.0505       pars:   0.26175 -0.07983 0.18359 -0.20922 -2.14184 0.37248       pars:   -0.23176 -0.61262 -0.36738 -0.54765 -2.27969 0.56084       pars:   -0.94983 -1.05323 0.06357 -0.34094 -2.11549 0.74746       pars:   -1.53946 -1.22704 -0.27451 -0.02058 -1.49041 0.86775       pars:   -1.08178 -1.10741 -0.1419 -0.36219 -2.01084 0.76743       pars:   -1.16952 -1.16077 -0.0098 -0.47559 -1.83181 0.72917       pars:   -1.2777 -1.21366 -0.03001 -0.29752 -1.67967 0.6613       pars:   -1.37196 -1.31467 -0.09354 -0.41658 -1.50044 0.59098       pars:   -1.47712 -1.38822 0.04369 -0.4013 -1.30894 0.54446       pars:   -1.60078 -1.42365 -0.04172 -0.37129 -1.08618 0.54342       pars:   -1.66212 -1.55665 -0.05341 -0.43194 -1.06783 0.32259       pars:   -1.60078 -1.42365 -0.04172 -0.37129 -1.08618 0.54342      
#>  pars:   1 1 0 0 -1.35922 0       pars:   0.27843 0.51119 0.15877 0.41388 -1.48525 0.16739       pars:   -1.88627 -0.95525 0.63506 1.65552 -1.86333 0.66956       pars:   -4.13528 -2.45781 -0.47532 -0.96337 -2.4054 1.21621       pars:   -2.057 -1.06723 0.39118 1.0681 -1.91932 0.71354       pars:   -2.32513 -1.2831 0.33201 0.50491 -1.99609 0.79248       pars:   -2.7642 -1.52612 0.46603 0.91439 -2.0378 0.90242       pars:   -3.27492 -1.74828 0.3391 0.59479 -2.04592 1.0534       pars:   -3.6356 -1.73671 0.86682 0.55539 -1.84337 1.05078       pars:   -3.30549 -1.74246 0.49429 0.71032 -2.01397 1.04118       pars:   -3.45476 -1.74957 0.4091 0.70481 -1.91425 1.04264       pars:   -3.55981 -1.79416 0.51279 0.63194 -1.81219 1.03447       pars:   -3.58609 -1.90374 0.49882 0.75005 -1.71519 0.97691       pars:   -3.67739 -1.96382 0.42195 0.66799 -1.59403 0.95971       pars:   -3.70249 -1.93003 0.4985 0.69558 -1.5565 0.95021       pars:   -3.70249 -1.93003 0.4985 0.69558 -1.5565 0.95021      

## called automatically within fit_ssm()
fit <- fit_ssm(sese2_n, date = "time", coord = c("longitude", "latitude"), 
tz = "America/Halifax", model = "crw", time.step = 24)
#> fitting crw SSM to 2 tracks...
#>  pars:   1 1 0 0 -1.82873 0       pars:   0.593 0.46862 -0.38936 -0.59345 -2.04237 0.0505       pars:   0.26175 -0.07983 0.18359 -0.20922 -2.14184 0.37248       pars:   -0.23176 -0.61262 -0.36738 -0.54765 -2.27969 0.56084       pars:   -0.94983 -1.05323 0.06357 -0.34094 -2.11549 0.74746       pars:   -1.53946 -1.22704 -0.27451 -0.02058 -1.49041 0.86775       pars:   -1.08178 -1.10741 -0.1419 -0.36219 -2.01084 0.76743       pars:   -1.16952 -1.16077 -0.0098 -0.47559 -1.83181 0.72917       pars:   -1.2777 -1.21366 -0.03001 -0.29752 -1.67967 0.6613       pars:   -1.37196 -1.31467 -0.09354 -0.41658 -1.50044 0.59098       pars:   -1.47712 -1.38822 0.04369 -0.4013 -1.30894 0.54446       pars:   -1.60078 -1.42365 -0.04172 -0.37129 -1.08618 0.54342       pars:   -1.66212 -1.55665 -0.05341 -0.43194 -1.06783 0.32259       pars:   -1.60078 -1.42365 -0.04172 -0.37129 -1.08618 0.54342      
#>  pars:   1 1 0 0 -1.35922 0       pars:   0.27843 0.51119 0.15877 0.41388 -1.48525 0.16739       pars:   -1.88627 -0.95525 0.63506 1.65552 -1.86333 0.66956       pars:   -4.13528 -2.45781 -0.47532 -0.96337 -2.4054 1.21621       pars:   -2.057 -1.06723 0.39118 1.0681 -1.91932 0.71354       pars:   -2.32513 -1.2831 0.33201 0.50491 -1.99609 0.79248       pars:   -2.7642 -1.52612 0.46603 0.91439 -2.0378 0.90242       pars:   -3.27492 -1.74828 0.3391 0.59479 -2.04592 1.0534       pars:   -3.6356 -1.73671 0.86682 0.55539 -1.84337 1.05078       pars:   -3.30549 -1.74246 0.49429 0.71032 -2.01397 1.04118       pars:   -3.45476 -1.74957 0.4091 0.70481 -1.91425 1.04264       pars:   -3.55981 -1.79416 0.51279 0.63194 -1.81219 1.03447       pars:   -3.58609 -1.90374 0.49882 0.75005 -1.71519 0.97691       pars:   -3.67739 -1.96382 0.42195 0.66799 -1.59403 0.95971       pars:   -3.70249 -1.93003 0.4985 0.69558 -1.5565 0.95021       pars:   -3.70249 -1.93003 0.4985 0.69558 -1.5565 0.95021