join an mpm-estimated behavioural index to ssm-predicted locations

join() joins ssm-predicted locations and mpm-estimated behavioural index into a single tibble. If the ssm-predicted tibble is a projected sf object then the output of join will also be an sf object (default). This can be avoided by using as_sf = FALSE.

join(
  ssm,
  mpm,
  what.ssm = "predicted",
  as_sf = FALSE,
  normalise = FALSE,
  group = FALSE
)

Arguments

ssm

an aniMotum ssm fitted model object

mpm

an aniMotum mpm fitted model object

what.ssm

specifies whether ssm predicted or fitted values are to be extracted

as_sf

logical; if FALSE then return a tibble with un-projected lonlat coordinates, otherwise return an sf tibble

normalise

logical; if output includes a move persistence estimate, should g (the move persistence index) be normalised to have minimum = 0 and maximum = 1 (default = FALSE).

group

logical; should g be normalised among individuals as a group, a 'relative g', or separately to highlight regions of lowest and highest move persistence along a track (default = FALSE).

Value

a single tbl with all individuals

Examples

## load example aniMotum fit objects (to save time)
## generate a ssm fit object
xs <- fit_ssm(ellie, spdf=FALSE, model = "rw", time.step=24, control = ssm_control(verbose = 0))
#> Error in nlminb(obj$par, ifelse(control$verbose == 1, myfn, obj$fn), obj$gr,  : 
#>   NA/NaN gradient evaluation
#> 
#> Error in updateCholesky(L, hessian.random) : 
#>   function 'sexp_as_cholmod_sparse' not provided by package 'Matrix'
#> Error in rep$pdHess: $ operator is invalid for atomic vectors
xm <- fit_mpm(xs, what = "p", model = "mpm")
#> fitting mpm...
#> Error in eval(expr, envir, enclos): object 'xs' not found

## join predicted values as an un-projected tibble
xsm <- join(xs, xm)
#> Error in eval(expr, envir, enclos): object 'xs' not found
xsm
#> Error in eval(expr, envir, enclos): object 'xsm' not found