Coerce input data into expected aniMotum format

format data by mapping supplied variable names to those expected by fit_ssm(), and ensuring variables are put into the expected order. Can be run manually by user as a data pre-processing step prior to calling fit_ssm() or can be called automatically by fit_ssm(). In the latter case, any custom variable names must be declared as arguments to fit_ssm(); see examples, below.

format_data(
  x,
  id = "id",
  date = "date",
  lc = "lc",
  coord = c("lon", "lat"),
  epar = c("smaj", "smin", "eor"),
  sderr = c("lonerr", "laterr"),
  tz = "UTC"
)

Arguments

x

input data

id

the name (as a quoted character string) of id variable: a unique identifier for individual (animal) track data sets.

date

the name (as a quoted character string)of the date/time variable: date and time (as YYYY-MM-DD HH:MM:SS) of each observation.

lc

the name (as a quoted character string) of the location quality class variable: Argos location quality class (values in the set: 3,2,1,0,"A","B","Z"). Can also include "G" for GPS data and/or "GL" for light-level geolocation (GLS) and other data types.

coord

the names (as quoted character strings) of the location coordinate variables: defaults are c("lon","lat"), but could also be c("x","y") for planar coordinates; or if input data is an sf object then "geometry". If input data is an sf object then coord is set to "geometry" by default.

epar

the names (as quoted character strings) of the Argos error ellipse parameters: defaults are "smaj" (ellipse semi-major axis), "smin" (ellipse semi-minor axis), and "eor" (ellipse orientation). Ignored if these variables are missing from the input data.

sderr

the names (as quoted character strings) of provided standard errors in longitude and latitude: defaults are "lonerr", "laterr". Typically, these are only provided for processed light-level geolocation data. Ignored if these variables are missing from the input data.

tz

the timezone the applies to the data/time variable if they are not in tz = 'UTC'. A list of valid timezone names can be viewed via OlsonNames()

Value

a data.frame or sf-tibble of input data in expected aniMotum format. Additional columns required by fit_ssm(), if missing, will be added to the formatted tibble: smaj, smin, eor, lonerr, and laterr.

Examples

## as a data pre-processing step
data(sese2_n)
head(sese2_n, 5)
#> # A tibble: 5 × 5
#>   longitude latitude time                lc    id       
#>       <dbl>    <dbl> <chr>               <fct> <chr>    
#> 1      72.5    -50.2 2009-02-01 17:50:46 A     ct36-E-09
#> 2      73.0    -50.4 2009-02-02 03:30:26 A     ct36-E-09
#> 3      73.8    -50.8 2009-02-02 17:50:48 B     ct36-E-09
#> 4      74.6    -51.2 2009-02-03 07:39:08 A     ct36-E-09
#> 5      74.9    -51.4 2009-02-03 15:29:59 A     ct36-E-09
d <- format_data(sese2_n, date = "time", coord = c("longitude","latitude"), 
tz = "America/Halifax")
fit <- fit_ssm(d, model = "crw", time.step = 24)
#> fitting crw SSM to 2 tracks...
#>  pars:   0 0 0 0       pars:   -0.66839 -0.32202 -0.66802 0.05745       pars:   -1.28283 0.26151 -0.17032 0.24251       pars:   -2.04068 -0.35231 -0.17829 0.46346       pars:   -1.40737 -0.07903 -0.40449 0.25551       pars:   -1.7747 0.14305 -0.39501 0.30181       pars:   -1.51575 -0.01351 -0.40169 0.26917       pars:   -1.58385 -0.11571 -0.36798 0.27295       pars:   -1.539 -0.05284 -0.39164 0.27038       pars:   -1.57736 -0.02627 -0.38806 0.2706       pars:   -1.57736 -0.02627 -0.38806 0.2706      
#>  pars:   0 0 0 0       pars:   -0.76745 0.06208 0.62625 0.1224       pars:   -2.40381 0.19445 1.96154 0.38338       pars:   -3.76779 0.7448 0.55152 0.7029       pars:   -2.19925 0.38788 0.59059 0.39942       pars:   -2.70977 -0.66522 0.6601 1.32664       pars:   -2.27021 0.33203 0.69998 0.44655       pars:   -2.39122 0.33328 0.70655 0.53403       pars:   -2.52103 0.33461 0.7136 0.62788       pars:   -2.39072 0.51956 0.86625 0.77449       pars:   -2.49192 0.36981 0.7443 0.65372       pars:   -2.48304 0.34004 0.71409 0.69638       pars:   -2.45684 0.36269 0.74075 0.73866       pars:   -2.45684 0.36269 0.74075 0.73866      

## called automatically within fit_ssm()
fit <- fit_ssm(sese2_n, date = "time", coord = c("longitude", "latitude"), 
tz = "America/Halifax", model = "crw", time.step = 24)
#> fitting crw SSM to 2 tracks...
#>  pars:   0 0 0 0       pars:   -0.66839 -0.32202 -0.66802 0.05745       pars:   -1.28283 0.26151 -0.17032 0.24251       pars:   -2.04068 -0.35231 -0.17829 0.46346       pars:   -1.40737 -0.07903 -0.40449 0.25551       pars:   -1.7747 0.14305 -0.39501 0.30181       pars:   -1.51575 -0.01351 -0.40169 0.26917       pars:   -1.58385 -0.11571 -0.36798 0.27295       pars:   -1.539 -0.05284 -0.39164 0.27038       pars:   -1.57736 -0.02627 -0.38806 0.2706       pars:   -1.57736 -0.02627 -0.38806 0.2706      
#>  pars:   0 0 0 0       pars:   -0.76745 0.06208 0.62625 0.1224       pars:   -2.40381 0.19445 1.96154 0.38338       pars:   -3.76779 0.7448 0.55152 0.7029       pars:   -2.19925 0.38788 0.59059 0.39942       pars:   -2.70977 -0.66522 0.6601 1.32664       pars:   -2.27021 0.33203 0.69998 0.44655       pars:   -2.39122 0.33328 0.70655 0.53403       pars:   -2.52103 0.33461 0.7136 0.62788       pars:   -2.39072 0.51956 0.86625 0.77449       pars:   -2.49192 0.36981 0.7443 0.65372       pars:   -2.48304 0.34004 0.71409 0.69638       pars:   -2.45684 0.36269 0.74075 0.73866       pars:   -2.45684 0.36269 0.74075 0.73866